requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.2
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.4
et-xmlfile==1.1.0
filelock==3.8.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
numpy==1.23.4
openai==0.23.1
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pandas-stubs==1.5.0.221012
pbr==5.10.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2==2.9.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sentry-sdk==1.9.10
six==1.16.0
sqlparse==0.4.3
stevedore==4.0.1
tqdm==4.64.1
types-pytz==2022.4.0.0
typing_extensions==4.4.0
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.5